requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pymysql==1.1.0
pydantic==2.9.2
pydantic-settings==2.5.2
#anthropic==0.34.0
# REPLACE with — includes Vertex AI support
#anthropic[vertex]==0.34.0
# FIX: upgrade anthropic to 0.40.0+ which uses httpx>=0.28.0 correctly
# 0.34.0 passes 'proxies' kwarg that httpx>=0.28.0 removed
anthropic[vertex]==0.40.0
# FIX: pin httpx to version compatible with anthropic 0.40.0
httpx==0.27.2
watchdog==5.0.3
python-dotenv==1.0.1
pandas==2.2.3
cryptography==43.0.1
